+++ Salveomensageiro/app/src/main/AndroidManifest.xml
@@ -2,8 +2,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
+
     <uses-permission android:name="android.permission.INTERNET" />
     <application
+
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -11,10 +13,9 @@
         android:label="@string/app_name"
         android:roundIcon="@drawable/salve2"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SalveOMensageiro"
-        tools:targetApi="31">
+        android:theme="@style/Theme.SalveOMensageiro">
         <activity
-            android:name=".ui.view.MainActivity"
+            android:name="br.com.salveomensageiro.ui.view.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.SalveOMensageiro.Splash">
             <intent-filter>
@@ -26,5 +27,6 @@
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
     </application>
+
 
 </manifest>
