+++ Salveomensageiro/app/src/main/java/br/com/salveomensageiro/data/repository/OrixaRepository.kt
@@ -1,0 +1,7 @@
+package br.com.salveomensageiro.data.repository
+
+import br.com.salveomensageiro.data.response.OrixaResponse
+
+interface OrixaRepository {
+    suspend fun getAllOrixas(): List<OrixaResponse>
+}
