+++ Salveomensageiro/app/src/main/java/br/com/salveomensageiro/ui/theme/Color.kt
@@ -1,0 +1,11 @@
+package br.com.salveomensageiro.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val RosaOrixa = Color(0xFFFFBF6F)
+val BackgroungCardBege = Color(0xFFF2E3CA)
+val DetailRedCard = Color(0xFFFF5349)
+
+val Purple40 = Color(0xFFF2C230)
+val PurpleGrey40 = Color(0xFFF2921D)
+val Pink40 = Color(0xFFFF5722)
