+++ Salveomensageiro/app/src/main/java/br/com/salveomensageiro/ui/view/OrixaState.kt
@@ -1,0 +1,11 @@
+package br.com.salveomensageiro.ui.view
+
+sealed class Screen(val route: String) {
+    object Home : Screen("orixasHome")
+    object Detail : Screen("detailCard/{index}") {
+        fun createRoute(rota: String) = "detailCard/{$rota}"
+    }
+
+    object Splash : Screen("splashScreen")
+
+}
