+++ Salveomensageiro/app/src/main/java/br/com/salveomensageiro/ui/view/SplashScreen.kt
@@ -1,0 +1,39 @@
+package br.com.salveomensageiro.ui.view
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.navigation.NavController
+import com.example.salveomensageiro.R
+import kotlinx.coroutines.delay
+
+@Composable
+fun SplashScreen(navController: NavController) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(MaterialTheme.colorScheme.primary),
+        contentAlignment = Alignment.Center
+    ) {
+        Image(
+            contentScale = ContentScale.Fit,
+            painter = painterResource(id = R.drawable.salve2),
+            contentDescription = "Logo da SplashScreen"
+        )
+        LaunchedEffect(key1 = true) {
+            delay(1000)
+            navController.navigate("orixasHome") {
+                popUpTo("splash_screen") { inclusive = true }
+            }
+        }
+    }
+}
