+++ Salveomensageiro/app/src/main/java/br/com/salveomensageiro/ui/viewmodel/OrixaViewmodelState.kt
@@ -1,0 +1,9 @@
+package br.com.salveomensageiro.ui.viewmodel
+
+import br.com.salveomensageiro.data.Orixa
+
+sealed class OrixaViewmodelState {
+    data class GetOrixas(val orixa: List<Orixa>) : OrixaViewmodelState()
+
+    object InitialState : OrixaViewmodelState()
+}
