+++ Salveomensageiro/app/src/main/java/br/com/salveomensageiro/ui/viewmodel/OrixasViewmodelFactory.kt
@@ -1,0 +1,12 @@
+package br.com.salveomensageiro.ui.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class OrixasViewmodelFactory : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+
+        return super.create(modelClass)
+    }
+
+}
