--- Salveomensageiro/app/src/main/java/com/example/salveomensageiro/ui/view/OrixaState.kt
@@ -1,11 +1,0 @@
-package com.example.salveomensageiro.ui.view
-
-sealed class Screen(val route: String) {
-    object Home : Screen("orixasHome")
-    object Detail : Screen("detailCard/{index}") {
-        fun createRoute(rota: String) = "detailCard/{$rota}"
-    }
-
-    object Splash : Screen("splashScreen")
-
-}
