--- Salveomensageiro/app/src/main/java/com/example/salveomensageiro/ui/viewmodel/OrixaViewmodelState.kt
@@ -1,9 +1,0 @@
-package com.example.salveomensageiro.ui.viewmodel
-
-import com.example.salveomensageiro.data.Orixa
-
-sealed class OrixaViewmodelState {
-    data class GetOrixas(val orixa: List<Orixa>) : OrixaViewmodelState()
-
-    object InitialState : OrixaViewmodelState()
-}
