--- Salveomensageiro/app/src/main/java/com/example/salveomensageiro/ui/viewmodel/OrixasViewmodelFactory.kt
@@ -1,13 +1,0 @@
-package com.example.salveomensageiro.ui.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.example.salveomensageiro.domain.OrixaRepositoryImpl
-
-class OrixasViewmodelFactory : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-
-        return super.create(modelClass)
-    }
-
-}
